create helper function for creating next app to reduce code repetition
